+++ cocoa-beans-common/src/test/java/net/apartium/cocoabeans/structs/MinecraftVersionTest.java
@@ -12,6 +12,7 @@ package net.apartium.cocoabeans.structs;
 
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
 import java.util.Objects;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -63,6 +64,8 @@ class MinecraftVersionTest {
 
         assertTrue(mySeverVersion.isHigherThan(MinecraftVersion.V1_21));
         assertFalse(MinecraftVersion.V1_21.isHigherThan(mySeverVersion));
+
+        assertTrue(MinecraftVersion.V26_1_1.isHigherThan(mySeverVersion));
     }
 
     @Test
@@ -100,6 +103,7 @@ class MinecraftVersionTest {
         assertTrue(mySeverVersion.isLowerThan(MinecraftVersion.V1_13));
         assertTrue(mySeverVersion.isLowerThan(MinecraftVersion.V1_13_1));
         assertTrue(mySeverVersion.isLowerThan(MinecraftVersion.V1_13_2));
+        assertTrue(mySeverVersion.isLowerThan(MinecraftVersion.V26_1));
 
         mySeverVersion = MinecraftVersion.V1_20_2;
 
@@ -128,6 +132,7 @@ class MinecraftVersionTest {
         assertFalse(mySeverVersion.isLowerThanOrEqual(MinecraftVersion.V1_12));
         assertFalse(mySeverVersion.isLowerThanOrEqual(MinecraftVersion.V1_9_4));
 
+        assertTrue(mySeverVersion.isLowerThanOrEqual(MinecraftVersion.V26_1));
         assertTrue(mySeverVersion.isLowerThanOrEqual(MinecraftVersion.V1_13));
         assertTrue(mySeverVersion.isLowerThanOrEqual(MinecraftVersion.V1_13_1));
         assertTrue(mySeverVersion.isLowerThanOrEqual(MinecraftVersion.V1_13_2));
@@ -166,6 +171,30 @@ class MinecraftVersionTest {
         assertEquals(0, MinecraftVersion.V1_18_1.compareTo(MinecraftVersion.V1_18_1));
         assertEquals(1, MinecraftVersion.V1_18_1.compareTo(MinecraftVersion.V1_9));
         assertEquals(-1, MinecraftVersion.V1_18_1.compareTo(MinecraftVersion.V1_20));
+        assertEquals(1, MinecraftVersion.V26_1.compareTo(MinecraftVersion.V1_9));
+
     }
 
+    @Test
+    void testToString() {
+        assertEquals("1.8", MinecraftVersion.V1_8.toString());
+        assertEquals("1.10.1", MinecraftVersion.V1_10_1.toString());
+        assertEquals("26.1", MinecraftVersion.V26_1.toString());
+        assertEquals("26.1.2", MinecraftVersion.V26_1_2.toString());
+    }
+
+    @Test
+    void getByProtocolVersion() {
+        assertEquals(List.of(MinecraftVersion.V1_8, MinecraftVersion.V1_8_1, MinecraftVersion.V1_8_2, MinecraftVersion.V1_8_3,
+                MinecraftVersion.V1_8_4, MinecraftVersion.V1_8_5, MinecraftVersion.V1_8_6, MinecraftVersion.V1_8_7,MinecraftVersion.V1_8_8,
+                MinecraftVersion.V1_8_9), MinecraftVersion.getByProtocolVersion(MinecraftVersion.V1_8.protocol()));
+
+        assertEquals(List.of(MinecraftVersion.V26_1, MinecraftVersion.V26_1_1, MinecraftVersion.V26_1_2), MinecraftVersion.getByProtocolVersion(MinecraftVersion.V26_1.protocol()));
+
+        assertEquals(List.of(), MinecraftVersion.getByProtocolVersion(4));
+        assertEquals(List.of(), MinecraftVersion.getByProtocolVersion(0));
+        assertEquals(List.of(), MinecraftVersion.getByProtocolVersion(500));
+        assertEquals(List.of(), MinecraftVersion.getByProtocolVersion(MinecraftVersionHolder.VERSIONS_BY_PROTOCOL.length));
+
+    }
 }